fix edit product

+++ controller/project.php
@@ -36,17 +36,17 @@ class ProjectController
 
         $project_id = Project::create($name, $contract, $address, $inn, $start_date, $end_date, $comment, $complaint, $zmo_id, $is_made_order, $document_scan, $documents, $is_ready);
         ProjectAccessController::create($project_id, json_encode(['all']), $_SESSION['user']['user_id']);
-        ProjectHistoryEditController::create('Создал проект ' . $project_id, null, null, $project_id, $_SESSION['user']['user_id']);
+        ProjectHistoryEditController::create('Создал проект ' . $project_id, null, null, 'create',$project_id, $_SESSION['user']['user_id']);
 
         if (!is_array($products[0])) return;
 
         for ($i = 0; $i < count($products[0]); $i++) {
             ProductController::create($products[0][$i], $products[1][$i], $products[2][$i], $products[3][$i], $products[4][$i], $products[5][$i], $products[6][$i], $products[7][$i], $products[8][$i], $products[9][$i], $products[10][$i], $project_id);
-            ProjectHistoryEditController::create('Добавил товар ' . $products[0][$i], null, null, $project_id, $_SESSION['user']['user_id']);
+            ProjectHistoryEditController::create('Добавил товар ' . $products[0][$i], null, null, 'add', $project_id, $_SESSION['user']['user_id']);
         }
     }
 
-    public static function edit($project_id, $name, $contract, $address, $inn, $start_date, $end_date, $comment, $complaint, $zmo_id, $is_made_order, $document_scan, $documents, $access_array)
+    public static function edit($project_id, $name, $contract, $address, $inn, $start_date, $end_date, $comment, $complaint, $zmo_id, $is_made_order, $document_scan, $documents, $products, $access_array)
     {
         $project_id = (int) $project_id;
         $name = DbQuery::protectedData($name);
@@ -65,7 +65,6 @@ class ProjectController
         $data_type = compact(
             'name',
             'contract',
-            'address',
             'inn',
             'start_date',
             'end_date',
@@ -108,10 +107,81 @@ class ProjectController
 
         if (!$query) return "Ошибка при изменении данных";
 
-        if (count($action_edit) < 1) return;
+        if (count($action_edit) > 1) {
+            foreach ($action_edit as $action) {
+                ProjectHistoryEditController::create('Изменил поле ' . $action['name'], $action['old'], $action['new'], 'edit', $project_id, $_SESSION['user']['user_id']);
+            }
+        }
 
-        foreach ($action_edit as $action) {
-            ProjectHistoryEditController::create($action['name'], $action['old'], $action['new'], $project_id, $_SESSION['user']['user_id']);
+        if (!is_array($products[0])) return;
+
+        for ($i = 0; $i < count($products[0]); $i++) {
+            $product_old = DbQuery::get('product', 'product_id', $products[11][$i])->fetch_assoc();
+            $product_new = $product_old;
+
+            if ($product_old) {
+                if (!ProjectController::checkEdited($product_old['name'], $products[0][$i])) {
+                    $product_new['name'] = $products[0][$i];
+                    ProjectHistoryEditController::create('Изменил поле name', $product_old['name'], $products[0][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['address_from'], $products[1][$i])) {
+                    $product_new['address_from'] = $products[1][$i];
+                    ProjectHistoryEditController::create('Изменил поле address_from', $product_old['address_from'], $products[1][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['address_to'], $products[2][$i])) {
+                    $product_new['address_to'] = $products[2][$i];
+                    ProjectHistoryEditController::create('Изменил поле address_to', $product_old['address_to'], $products[2][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['count'], $products[3][$i])) {
+                    $product_new['count'] = $products[3][$i];
+                    ProjectHistoryEditController::create('Изменил поле count', $product_old['count'], $products[3][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['unit_measurement'], $products[4][$i])) {
+                    $product_new['unit_measurement'] = $products[4][$i];
+                    ProjectHistoryEditController::create('Изменил поле unit_measurement', $product_old['unit_measurement'], $products[4][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['price'], $products[5][$i])) {
+                    $product_new['price'] = $products[5][$i];
+                    ProjectHistoryEditController::create('Изменил поле price', $product_old['price'], $products[5][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['amount'], $products[6][$i])) {
+                    $product_new['amount'] = $products[6][$i];
+                    ProjectHistoryEditController::create('Изменил поле amount', $product_old['amount'], $products[6][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['purchase_price'], $products[7][$i])) {
+                    $product_new['purchase_price'] = $products[7][$i];
+                    ProjectHistoryEditController::create('Изменил поле purchase_price', $product_old['purchase_price'], $products[7][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['purchase_amount'], $products[8][$i])) {
+                    $product_new['purchase_amount'] = $products[8][$i];
+                    ProjectHistoryEditController::create('Изменил поле purchase_amount', $product_old['purchase_amount'], $products[8][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['status_delivery'], $products[9][$i])) {
+                    $product_new['status_delivery'] = $products[9][$i];
+                    ProjectHistoryEditController::create('Изменил поле status_delivery', $product_old['status_delivery'], $products[9][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                if (!ProjectController::checkEdited($product_old['status_payment'], $products[10][$i])) {
+                    $product_new['status_payment'] = $products[10][$i];
+                    ProjectHistoryEditController::create('Изменил поле status_payment', $product_old['status_payment'], $products[10][$i], 'edit', $project_id, $_SESSION['user']['user_id']);
+                }
+
+                ProductController::update($product_new['product_id'], $product_new['name'], $product_new['address_from'], $product_new['address_to'], $product_new['count'], $product_new['unit_measurement'], $product_new['price'], $product_new['amount'], $product_new['purchase_price'], $product_new['purchase_amount'], $product_new['status_delivery'], $product_new['status_payment']);
+
+                continue;
+            }
+
+            ProductController::create($products[0][$i], $products[1][$i], $products[2][$i], $products[3][$i], $products[4][$i], $products[5][$i], $products[6][$i], $products[7][$i], $products[8][$i], $products[9][$i], $products[10][$i], $project_id);
+            ProjectHistoryEditController::create('Добавил товар ' . $products[0][$i], null, null, 'add', $project_id, $_SESSION['user']['user_id']);
         }
 
         header('Refresh: 0');
@@ -157,13 +227,14 @@ class ProjectController
 
         $user_power = (int) DbQuery::parse('role', 'role_id', $_SESSION['user']['role_id'], 'power');
 
-        if ($access_array[0] != 'all' || $user_power < 20) return "Ошибка при удалении";
+        if ($access_array[0] != 'all' || $user_power < 20 || array_search('project_delete', $access_array) !== false) return "Ошибка при удалении";
 
         $query = Project::delete($project_id);
 
         if (!$query) return "Ошибка при удалении";
 
-        ProjectHistoryEditController::create('Удалил проект ' . $project_id, null, null, $project_id, $_SESSION['user']['user_id']);
+        ProjectHistoryEditController::create('Удалил проект ' . $project_id, null, null, 'delete', $project_id, $_SESSION['user']['user_id']);
+        header('Location: /');
     }
 }
